Update NCP timeout type to uint16_t and clean up includes in SiWx NCP utility

+++ src/platform/silabs/wifi/SiWx/ncp/efx32_ncp_host.c
@@ -50,7 +50,7 @@
 
 #define MAX_DATA_PACKET_SIZE 1800
 
-static const uint8_t ncp_transfer_timeout_ms = 1000;
+static const uint16_t ncp_transfer_timeout_ms = 1000;
 
 // use SPI handle for EXP header (configured in project settings)
 extern SPIDRV_Handle_t sl_spidrv_exp_handle;
